# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Spend Little",
        "description": "Essa API foi desenvolvida para manipular Spend Little",
        "contact": {
            "email": "gezerramo@gmail.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://spend-little.herokuapp.com/api/v1",
            "description": "API(Produção)"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API(Produção)"
        }
    ],
    "paths": {
        "/account/me": {
            "get": {
                "tags": [
                    "/account"
                ],
                "summary": "Account",
                "description": "Responsible route recover account data!!",
                "operationId": "228c02411f7bff10c6d4197280066e8c",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "/account"
                ],
                "summary": "Account",
                "description": "Rota responsavel por editar conta do usuario logado!",
                "operationId": "d375f0c480fcb91943e1fb98e822f11d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Fulano de Tal",
                                    "email": "fulano@gmail.com",
                                    "password": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/account/refresh": {
            "get": {
                "tags": [
                    "/account"
                ],
                "summary": "Account",
                "description": "Responsible route refresh token!",
                "operationId": "5de51af5f3f2e5d9432d86d22610c50a",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/account/me/image": {
            "post": {
                "tags": [
                    "/account"
                ],
                "summary": "Account",
                "description": "Rota responsavel por editar conta do usuario logado!",
                "operationId": "38bbe3d69564c7a1f69342572b5a0db2",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "imagem": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}